Fixed compiler warnings when compiling for 64 bit

+++ src/JSTTextView.m
@@ -373,7 +373,7 @@ static void initBraces() {
         
         if (!openingBraces) {
             unichar charBuf[100];
-            unsigned defLen;
+            NSUInteger defLen;
             
             defLen = [defaultOpeningBraces length];
             [defaultOpeningBraces getCharacters:charBuf];
